creates query script file and main menu prompts
